requirements.txt
altgraph==0.17.4
certifi==2025.10.5
charset-normalizer==3.4.4
et_xmlfile==2.0.0
idna==3.11
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pefile==2023.2.7
pillow==12.0.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
requests==2.32.5
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
bcrypt==5.0.0